refactor: move speech handler
