+++ handlers/speech.py
